CRUD AdminInputs
- Make separate create and update forms and views for each AdminInput
type. Couldn't handle it in one view; the way we have set up the
database would mean too much work if a user would try to change
input_type, when updating an AdminInput.
- Display DropDownAdminInput choices attribute as a TextArea field
in the form. Convert it to a JSON.
- Manage created_by AdminInput Field. When a user creates a new object,
set that user as the value of created_by

See #71